+++ src/parser/parser.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <utility>
 #include <vector>
 #include <assert.h>
@@ -22,13 +23,13 @@ public:
 private:
 	Lexer *lexer;
 	std::vector<Token> tokens;
-	long unsigned int i = 0;
+	uint64_t i = 0;
 	int function_addr = 0;
 	std::vector<std::string> include_paths;
 
 	bool is_legal_name(Token token_name);
 	std::string add_escapes_to_string(std::string str);
-	long long eval_const_expression(Location definition_loc);
+	int64_t eval_const_expression(Location definition_loc);
 
 	Op convert_token_to_op(Token tok, std::map<std::string, std::pair<LangType, int>> var_offsets = {});
 	std::vector<Op> link_ops(std::vector<Op> ops, std::map<std::string, std::pair<int, int>> labels);
